Refactored omniglot data on its own file
