+++ few_shot/dataset/__init__.py
@@ -1,34 +1,9 @@
-import os
-
 from typing import Callable, List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
 import sklearn.preprocessing as skp
 import tensorflow as tf
-
-
-def create_omniglot_df(path: str) -> pd.DataFrame:
-    """Creates a DataFrame for Omniglot data based on its directory structure.
-
-    :param path: path to the base folder of the processed Omniglot dataset.
-    """
-    samples = []
-    for dirpath, dirnames, filenames in os.walk(path):
-        if not filenames:
-            continue
-        *_, alphabet, character = dirpath.split(os.path.sep)
-        class_name = f'{alphabet}.{character}'
-        for f in filenames:
-            samples.append(
-                {
-                    'class_name': class_name,
-                    'alphabet': alphabet,
-                    'character': character,
-                    'filepath': os.path.abspath(os.path.join(dirpath, f))
-                })
-
-    return pd.DataFrame.from_records(samples)
 
 
 class FewShotEpisodeGenerator(object):
